Migration: Split style folder in components and pages
