+++ new-web/src/style/components/ControlButtons.css
