+++ new-web/src/style/components/DocumentControlButtons.css
